move BOM unit tests into maven module.
