+++ google-cloud-apis/testSrc/com/google/cloud/tools/intellij/cloudapis/GoogleCloudApiSelectorPanelTest.java
@@ -17,16 +17,12 @@
 package com.google.cloud.tools.intellij.cloudapis;
 
 import static com.google.common.truth.Truth.assertThat;
-import static org.mockito.Mockito.when;
 
-import com.google.cloud.tools.intellij.GctFeature;
 import com.google.cloud.tools.intellij.cloudapis.maven.CloudApiMavenService;
-import com.google.cloud.tools.intellij.cloudapis.maven.CloudLibraryMavenProjectState;
 import com.google.cloud.tools.intellij.project.CloudProject;
 import com.google.cloud.tools.intellij.project.ProjectSelector;
 import com.google.cloud.tools.intellij.service.PluginInfoService;
 import com.google.cloud.tools.intellij.testing.CloudToolsRule;
-import com.google.cloud.tools.intellij.testing.MavenTestUtils;
 import com.google.cloud.tools.intellij.testing.TestFixture;
 import com.google.cloud.tools.intellij.testing.TestModule;
 import com.google.cloud.tools.intellij.testing.TestService;
@@ -36,35 +32,24 @@ import com.google.cloud.tools.intellij.testing.apis.TestCloudLibrary.TestCloudLi
 import com.google.cloud.tools.libraries.json.CloudLibrary;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Lists;
 import com.intellij.openapi.application.ApplicationManager;
-import com.intellij.openapi.application.Result;
-import com.intellij.openapi.command.WriteCommandAction;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.module.ModuleManager;
 import com.intellij.testFramework.ThreadTracker;
 import com.intellij.testFramework.fixtures.IdeaProjectTestFixture;
-import com.intellij.util.xml.DomUtil;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import javax.swing.JCheckBox;
-import javax.swing.JComboBox;
 import javax.swing.JPanel;
 import javax.swing.JTable;
 import javax.swing.table.TableModel;
 import org.eclipse.aether.version.Version;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.idea.maven.dom.MavenDomUtil;
-import org.jetbrains.idea.maven.dom.model.MavenDomDependency;
-import org.jetbrains.idea.maven.dom.model.MavenDomProjectModel;
-import org.jetbrains.idea.maven.project.MavenProject;
-import org.jetbrains.idea.maven.project.MavenProjectsManager;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -269,121 +254,6 @@ public final class GoogleCloudApiSelectorPanelTest {
     assertThat((Boolean) model.getValueAt(1, 1)).isTrue();
 
     assertDetailsShownForLibrary(panel.getDetailsPanel(), LIBRARY_1, JAVA_CLIENT_1);
-  }
-
-  @Test
-  public void getPanel_withAvailableBomVersions_populatesBomVersionsInReverseOrder() {
-    // TODO (eshaul): remove once feature is released
-    when(pluginInfoService.shouldEnable(GctFeature.BOM)).thenReturn(true);
-
-    when(mavenService.getAllBomVersions()).thenReturn(ImmutableList.of("v0", "v1", "v2"));
-
-    GoogleCloudApiSelectorPanel panel =
-        new GoogleCloudApiSelectorPanel(ImmutableList.of(), testFixture.getProject());
-
-    JComboBox<String> bomComboBox = panel.getBomComboBox();
-
-    assertThat(panel.getBomSelectorLabel().isVisible()).isTrue();
-    assertThat(panel.getBomComboBox().isVisible()).isTrue();
-    assertThat(bomComboBox.getItemCount()).isEqualTo(3);
-
-    assertThat(bomComboBox.getItemAt(0)).isEqualTo("v2");
-    assertThat(bomComboBox.getItemAt(1)).isEqualTo("v1");
-    assertThat(bomComboBox.getItemAt(2)).isEqualTo("v0");
-  }
-
-  @Test
-  public void getPanel_withManyAvailableBomVersions_limitsNumBomVersions() {
-    // TODO (eshaul): remove once feature is released
-    when(pluginInfoService.shouldEnable(GctFeature.BOM)).thenReturn(true);
-
-    List<String> versions = Lists.newArrayList();
-    for (int i = 0; i < 20; i++) {
-      versions.add("v" + i);
-    }
-
-    when(mavenService.getAllBomVersions()).thenReturn(versions);
-
-    GoogleCloudApiSelectorPanel panel =
-        new GoogleCloudApiSelectorPanel(ImmutableList.of(), testFixture.getProject());
-
-    assertThat(panel.getBomComboBox().getItemCount()).isEqualTo(5);
-  }
-
-  @Test
-  public void getPanel_withNoAvailableBomVersions_hidesBomUi() {
-    // TODO (eshaul): remove once feature is released
-    when(pluginInfoService.shouldEnable(GctFeature.BOM)).thenReturn(true);
-
-    when(mavenService.getAllBomVersions()).thenReturn(ImmutableList.of());
-
-    GoogleCloudApiSelectorPanel panel =
-        new GoogleCloudApiSelectorPanel(ImmutableList.of(), testFixture.getProject());
-
-    assertThat(panel.getBomSelectorLabel().isVisible()).isFalse();
-    assertThat(panel.getBomComboBox().isVisible()).isFalse();
-  }
-
-  @Test
-  public void getPanel_withBomInPom_partOfAvailableBoms_preselectsConfiguredVersion() {
-    // TODO (eshaul): remove once feature is released
-    when(pluginInfoService.shouldEnable(GctFeature.BOM)).thenReturn(true);
-
-    when(mavenService.getAllBomVersions()).thenReturn(ImmutableList.of("v0", "v1", "v2"));
-
-    MavenTestUtils.getInstance()
-        .runWithMavenModule(
-            testFixture.getProject(),
-            module -> {
-              String preconfigureBomVersion = "v1";
-              writeBomDependency(module, preconfigureBomVersion);
-              CloudLibraryMavenProjectState.getInstance(testFixture.getProject())
-                  .syncCloudLibrariesBom();
-
-              GoogleCloudApiSelectorPanel panel =
-                  new GoogleCloudApiSelectorPanel(ImmutableList.of(), testFixture.getProject());
-
-              // Set the selected module to the one with the preconfigured BOM
-              panel.getModulesComboBox().setSelectedItem(module);
-
-              JComboBox<String> bomComboBox = panel.getBomComboBox();
-              assertThat(bomComboBox.getItemCount()).isEqualTo(3);
-              assertThat(bomComboBox.getItemAt(0)).isEqualTo("v2");
-              assertThat(bomComboBox.getItemAt(1)).isEqualTo("v1");
-              assertThat(bomComboBox.getItemAt(2)).isEqualTo("v0");
-
-              assertThat(bomComboBox.getSelectedItem()).isEqualTo(preconfigureBomVersion);
-            });
-  }
-
-  @Test
-  public void getPanel_withBomInPom_withNoAvailableBoms_hasOnlyPreconfiguredBom() {
-    // TODO (eshaul): remove once feature is released
-    when(pluginInfoService.shouldEnable(GctFeature.BOM)).thenReturn(true);
-
-    when(mavenService.getAllBomVersions()).thenReturn(ImmutableList.of());
-
-    MavenTestUtils.getInstance()
-        .runWithMavenModule(
-            testFixture.getProject(),
-            module -> {
-              String preconfigureBomVersion = "v1";
-              writeBomDependency(module, preconfigureBomVersion);
-              CloudLibraryMavenProjectState.getInstance(testFixture.getProject())
-                  .syncCloudLibrariesBom();
-
-              GoogleCloudApiSelectorPanel panel =
-                  new GoogleCloudApiSelectorPanel(ImmutableList.of(), testFixture.getProject());
-
-              // Set the selected module to the one with the preconfigured BOM
-              panel.getModulesComboBox().setSelectedItem(module);
-
-              JComboBox<String> bomComboBox = panel.getBomComboBox();
-              assertThat(bomComboBox.getItemCount()).isEqualTo(1);
-              assertThat(bomComboBox.getItemAt(0)).isEqualTo("v1");
-
-              assertThat(bomComboBox.getSelectedItem()).isEqualTo(preconfigureBomVersion);
-            });
   }
 
   @Test
@@ -685,27 +555,6 @@ public final class GoogleCloudApiSelectorPanelTest {
     // TODO(nkibler): Consider refactoring the details panel to allow unit tests to inject fake
     // icons. Until then, this will always be null.
     assertThat(panel.getIcon().getIcon()).isNull();
-  }
-
-  private void writeBomDependency(Module module, String bomVersion) {
-    MavenProject mavenProject =
-        MavenProjectsManager.getInstance(testFixture.getProject()).findProject(module);
-    MavenDomProjectModel model =
-        MavenDomUtil.getMavenDomProjectModel(testFixture.getProject(), mavenProject.getFile());
-
-    new WriteCommandAction(testFixture.getProject(), DomUtil.getFile(model)) {
-      @Override
-      protected void run(@NotNull Result result) {
-        MavenDomDependency bomDependency =
-            model.getDependencyManagement().getDependencies().addDependency();
-
-        bomDependency.getGroupId().setStringValue(CloudApiMavenService.GOOGLE_CLOUD_JAVA_BOM_GROUP);
-        bomDependency
-            .getArtifactId()
-            .setStringValue(CloudApiMavenService.GOOGLE_CLOUD_JAVA_BOM_ARTIFACT);
-        bomDependency.getVersion().setStringValue(bomVersion);
-      }
-    }.execute();
   }
 
   /**
